:construction: Initial tests for `pickle` support
